:bug: Fix long/integerfield validation on write and add regression test

+++ backend/src/zac/camunda/dynamic_forms/serializers.py
@@ -57,7 +57,7 @@ class FormFieldSerializer(PolymorphicSerializer):
     serializer_mapping = {
         "enum": EnumSerializer,
         "string": StringSerializer,
-        "int": IntSerializer,
+        "long": IntSerializer,
         "boolean": BooleanSerializer,
         "date": DatetimeSerializer,
     }
